Created standard class for buttons, file for card help functions and a new style property on the image, created CSS file for welcome and part of the stylization of the login page
