+++ src/features/components/Image.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Image = ({ src, alt, width, className, }) => (
-  <img src={src} alt={alt} width={width} className={className} />
+const Image = ({ src, alt, width, className, style, }) => (
+  <img src={src} alt={alt} width={width} className={className} style={style} />
 );
 
 export default Image;
@@ -10,6 +10,7 @@ export default Image;
 Image.defaultProps = {
   width: `${100}%`,
   className: '',
+  style: {},
 };
 
 Image.propTypes = {
@@ -17,4 +18,5 @@ Image.propTypes = {
   src: PropTypes.string.isRequired,
   alt: PropTypes.string.isRequired,
   className: PropTypes.string.isRequired,
+  style: PropTypes.objectOf(Object).isRequired,
 };
