+++ src/image/index.jsx
@@ -5,6 +5,8 @@ export { default as Passo3 } from './perguntas-plataforma.png';
 export { default as Passo4 } from './geracao-etiquetas.png';
 export { default as Passo5 } from './excecoes-logisticas.png';
 export { default as Passo6 } from './dashboard-operacional.png';
-export { default as Passo7 } from './integracao-nota-fiscal.png';
-// export { default as Passo8 } from './';
+export { default as Passo7 } from './oferta-inteligente.png';
+export { default as Passo8 } from './integracao-nota-fiscal.png';
+
+export { default as Engatinhando } from './olistinho-engatinhando.png';
