+++ src/index.css
@@ -1,3 +1,7 @@
+html {
+  background-color: #fbfbfb;
+}
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
@@ -10,4 +14,14 @@ body {
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
+}
+
+.btn {
+  color: #ffffff;
+  font-weight: 400;
+  text-align: center;
+  cursor: pointer;
+  padding: 6px 12px;
+  font-size: 14px;
+  border: 0;
 }
