+++ src/pages/Index.js
@@ -1,0 +1,4 @@
+export { default as Dashboard } from './Dashboard';
+export { default as Login } from './Login';
+export { default as Welcome } from './Welcome';
+export { default as Order } from './Order';
